+++ CHANGES.rst
@@ -5,6 +5,7 @@ not released
 ------------
 
 - The ``filtering.rpsl`` section of general and clients configuration files has been renamed into ``filtering.irrdb``.
+- The command line argument ``--template-dir`` has been renamed into ``--templates-dir``.
 
 v0.1.0a10
 ---------